requirements.txt
SQLAlchemy==2.0.31
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
pytest==8.2.2
python-dotenv==1.0.1
python-multipart==0.0.9
psycopg2-binary